gestion de l'afichage du status

+++ src/regub/commercial/Video.java
@@ -154,7 +154,7 @@ public class Video {
         this.date_validation.set(dateval);
     }
 
-    public double getStatut() {
+    public int getStatut() {
         return statut.get();
     }
